// src/app/modules/asegurado/models/asegurado.ts
export interface Asegurado {
  personaId: number
  aseguradoId: number
  nombres?: string
  paterno?: string
  materno?: string
  fechaNacimiento?: string
  documentoIdentidad?: string
  sexo: number
  matricula?: string
  parametroTipoParentescoId?: string
  parametroEstadoId: number
  estadoAsegurado?: string
  estadoTitularId: number
  estadoTitular?: string
  parametroTipoSeguroId: number
  tipoSeguro?: string
  ubicacionRegionalId: number
  ubicacion?: string
  razonSocial?: string
  nit?: string
  nroPatronal?: string
  centroMedico?: string
  titular?: string
  documentoIdentidadTitular?: string
  matriculaTitular?: string
  imagen?: string
  parentesco?: string
  isValidado?: true
  centroMedicoId: number
  consultorioId: number
  consultorio?: string
  direccion?: string
  estadoAporteId: number
  estadoAporte?: string
  ultimoAporte?: string
  isRevisado?: string
  celular?: string
  emailPersona?: string
  emailEmpresa?: string
  empresaAfiliadaId: number
  empresaId: number
  codigo?: string
  tipoRegistro: number
  estadoRegistroId: number
  observacionRegulacion?: string
  estadoMora?: string
  estadoEmpresa?: string
  grupoFamiliarId: number
  fechaInicioLaboral?: Date
  fechaFinLaboral: Date
  parametroEstadoEmpresaId: number
  fechaAfiliacion: Date
  fechaNacimientoTitular: Date
  parametroTipoBeneficiario: number
  fechaFinAfiliacion: Date
  tasa: number
  parametroSectorId: number
  sector?: string
  pareja?: string
  descripcionGenero?: string
}