// src/app/modules/asegurado/models/grupo-familiar.ts
export interface GrupoFamiliar {
  aseguradoId: number
  personaId: number
  matricula: string
  documentoIdentidad: number
  extension: string
  paterno: string
  materno: string
  nombres: string
  fechaNacimiento: Date
  parentesco: string
  parametroTipoParentescoId: number
  paciente: string
  parametroEstadoId: string
  isValidado: number
  info: number
  infoVal2: string
}