// src/app/modules/asegurado/models/regularizacion.ts
export interface Regularizacion {
  regularizacionId: number
  estado: number
  tipoModificacion: number
  nroForm: string
  primerApellido: string
  segundoApellido: string
  extension: string
  nombres: string
  documentoIdentidad: string
  complemento: string
  fechaNacimiento: Date
  genero: number
  tipoBeneficiario: number
  beneficiarioId: number
  descripcionParentesco: string
  descripcionLugarNacimiento: string
  datos: string
}